+++ CHANGELOG.md
@@ -1,5 +1,5 @@
 # 7.0.0-beta.5
-- [#33 Update GOV.UK Frontend to 0.0.27](https://github.com/alphagov/govuk-prototype-kit-private-beta/pull/33)
+- [#36 Update GOV.UK Frontend to 0.0.28](https://github.com/alphagov/govuk-prototype-kit-private-beta/pull/36)
 
 # 7.0.0-beta.4