+++ CHANGELOG.md
@@ -1,4 +1,4 @@
-# Unreleased
+# 7.0.0-beta.6
 
 Breaking changes: